// packages/components/src/cli/theming/tokens/color/color.types.ts
export type Name =
  | "foregroundNeutral"
  | "foregroundNeutralFaded"
  | "foregroundDisabled"
  | "foregroundPrimary"
  | "foregroundCritical"
  | "foregroundPositive"
  | "borderNeutral"
  | "borderNeutralFaded"
  | "borderDisabled"
  | "borderPrimary"
  | "borderPrimaryFaded"
  | "borderCritical"
  | "borderCriticalFaded"
  | "borderPositive"
  | "borderPositiveFaded"
  | "backgroundNeutral"
  | "backgroundNeutralFaded"
  | "backgroundDisabled"
  | "backgroundDisabledFaded"
  | "backgroundPrimary"
  | "backgroundPrimaryFaded"
  | "backgroundCritical"
  | "backgroundCriticalFaded"
  | "backgroundPositive"
  | "backgroundPositiveFaded"
  | "backgroundNeutralHighlighted"
  | "backgroundPrimaryHighlighted"
  | "backgroundCriticalHighlighted"
  | "backgroundPositiveHighlighted"
  | "backgroundPage"
  | "backgroundPageFaded"
  | "backgroundBase"
  | "backgroundElevated"
  | "white"
  | "black";
export type GeneratedOnName =
  | "onBackgroundNeutral"
  | "onBackgroundPrimary"
  | "onBackgroundPositive"
  | "onBackgroundCritical"
  | "onBackgroundNeutralHighlighted"
  | "onBackgroundPrimaryHighlighted"
  | "onBackgroundPositiveHighlighted"
  | "onBackgroundCriticalHighlighted";
export type GeneratedRGBName =
  | "rgbBackgroundNeutral"
  | "rgbBackgroundNeutralFaded"
  | "rgbBackgroundDisabled"
  | "rgbBackgroundDisabledFaded"
  | "rgbBackgroundPrimary"
  | "rgbBackgroundPrimaryFaded"
  | "rgbBackgroundCritical"
  | "rgbBackgroundCriticalFaded"
  | "rgbBackgroundPositive"
  | "rgbBackgroundPositiveFaded"
  | "rgbBackgroundNeutralHighlighted"
  | "rgbBackgroundPrimaryHighlighted"
  | "rgbBackgroundCriticalHighlighted"
  | "rgbBackgroundPage"
  | "rgbBackgroundPageFaded"
  | "rgbBackgroundBase"
  | "rgbBackgroundElevated";

export type Token = { hex: string; hexDark?: string };